+++ changelog.md
@@ -1,3 +1,17 @@
+# v2.3.3
+##### Added
+- **IM Person Proxy**
+	- new proxy to handle person operation
+	- new method `searchPersonFromRoot (session)`
+	- new method `searchPersonFromRoot (session, filter)`
+	- new method `lookupPerson (session, dn)`
+##### Changed
+- **Examples**
+	- new funcionts to test `searchPersonFromRoot` and `lookupPerson` methods
+##### Removed
+- N/A
+
+
 # v2.3.2
 ##### Added
 - **IM Role Proxy**